Rename package to weblocalstorage
