+++ app/src/main/java/com/duckduckgo/app/browser/weblocalstorage/WebLocalStorageSettingsJsonParser.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.app.browser.localstorage
+package com.duckduckgo.app.browser.weblocalstorage
 
 import com.duckduckgo.common.utils.DispatcherProvider
 import com.duckduckgo.di.scopes.AppScope
