+++ app/src/test/java/com/duckduckgo/app/browser/weblocalstorage/WebLocalStorageSettingsJsonParserImplTest.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.app.browser.localstorage
+package com.duckduckgo.app.browser.weblocalstorage
 
 import com.duckduckgo.common.test.CoroutineTestRule
 import com.duckduckgo.common.test.FileUtilities
